@@ -7,17 +7,17 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-intl": "^4.7.6",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
-    "react-transition-group": "^4.4.1",
+    "react-redux": "^7.2.8",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.0",
+    "react-transition-group": "^4.4.2",
     "redux": "^4.0.5",
     "redux-saga": "^1.1.3",
-    "reselect": "^4.0.0",
-    "styled-components": "^5.2.1"
+    "reselect": "^4.1.5",
+    "styled-components": "^5.3.5"
   },
   "scripts": {
     "start": "react-scripts start",
