@@ -1,18 +1,13 @@
 /**
+ *
  * AppContainer
+ *
  */
-import React from "react";
-import { BrowserRouter } from "react-router-dom";
-
 // import containers
 import AppRoutes from "./AppRoutes";
 
 const AppContainer = () => {
-  return (
-    <BrowserRouter>
-      <AppRoutes />
-    </BrowserRouter>
-  );
+  return <AppRoutes />;
 };
 
 export default AppContainer;
