@@ -1,8 +1,8 @@
 /**
+ *
  * NotFoundContainer
+ *
  */
-import React from "react";
-
 // import components
 import NotFound from "../../components/NotFound";
 
