@@ -1,8 +1,8 @@
 /**
+ *
  * WelcomeContainer
+ *
  */
-import React from "react";
-
 // import components
 import Welcome from "../../components/Welcome";
 
