@@ -1,8 +1,8 @@
 /**
+ *
  * NotFoundPage
+ *
  */
-import React from "react";
-
 // import containers
 import NavigationContainer from "../../containers/NavigationContainer";
 import NotFoundContainer from "../../containers/NotFoundContainer";
